+++ packages/web/CHANGELOG.md
@@ -3,6 +3,42 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+# [3.24.0-alpha.0](https://github.com/standardnotes/app/compare/@standardnotes/web@3.23.0...@standardnotes/web@3.24.0-alpha.0) (2022-06-20)
+
+### Features
+
+* add "Files Navigation" experimental feature to Labs ([#1125](https://github.com/standardnotes/app/issues/1125)) ([b6cd4bf](https://github.com/standardnotes/app/commit/b6cd4bf6bf158d98f74f0221ab7d1271d1791982))
+
+# [3.23.0](https://github.com/standardnotes/app/compare/@standardnotes/web@3.23.0-alpha.1...@standardnotes/web@3.23.0) (2022-06-20)
+
+**Note:** Version bump only for package @standardnotes/web
+
+# [3.23.0-alpha.1](https://github.com/standardnotes/app/compare/@standardnotes/web@3.23.0-alpha.0...@standardnotes/web@3.23.0-alpha.1) (2022-06-20)
+
+**Note:** Version bump only for package @standardnotes/web
+
+# [3.23.0-alpha.0](https://github.com/standardnotes/app/compare/@standardnotes/web@3.22.7-alpha.0...@standardnotes/web@3.23.0-alpha.0) (2022-06-19)
+
+### Features
+
+* ctrl+a to select all items ([#1123](https://github.com/standardnotes/app/issues/1123)) ([dcf3724](https://github.com/standardnotes/app/commit/dcf3724e2c951d7bbad276d37b3302c34d7ee78f))
+
+## [3.22.7-alpha.0](https://github.com/standardnotes/app/compare/@standardnotes/web@3.22.6...@standardnotes/web@3.22.7-alpha.0) (2022-06-18)
+
+### Bug Fixes
+
+* exporting multiple notes with same title ([#1119](https://github.com/standardnotes/app/issues/1119)) ([dfbd72b](https://github.com/standardnotes/app/commit/dfbd72b1dfd4a4681c452c9113248045e69880cd))
+
+## [3.22.6](https://github.com/standardnotes/app/compare/@standardnotes/web@3.22.6-alpha.0...@standardnotes/web@3.22.6) (2022-06-18)
+
+**Note:** Version bump only for package @standardnotes/web
+
+## [3.22.6-alpha.0](https://github.com/standardnotes/app/compare/@standardnotes/web@3.22.5...@standardnotes/web@3.22.6-alpha.0) (2022-06-18)
+
+### Bug Fixes
+
+* plus editor icons ([#1120](https://github.com/standardnotes/app/issues/1120)) ([ba65948](https://github.com/standardnotes/app/commit/ba65948364a3fca7bfa5005c56802102c73ccd99))
+
 ## 3.22.5 (2022-06-16)
 
 **Note:** Version bump only for package @standardnotes/web